@@ -2,27 +2,27 @@ source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 ruby '2.7.1'
-gem 'rails', '4.2.11.3'
+gem 'rails', '6.1'
 
-gem 'sass-rails', '~> 6.0'
-gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.0.0'
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
 gem 'jquery-rails'
 gem 'turbolinks'
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder'
 gem 'sdoc', '~> 0.4.0',          group: :doc
 gem 'spring',        group: :development
 gem 'sprockets-rails', :require => 'sprockets/railtie'
 gem 'bootstrap', '~> 4.6'
 gem 'bourbon'
-gem 'devise',           '~> 4.2'
-gem 'devise_invitable', '~> 1.5', '>= 1.5.5'
+gem 'devise'
+gem 'devise_invitable'
 gem 'haml-rails'
 gem 'high_voltage'
-gem 'mysql2', '~> 0.5.3'
-gem 'font-awesome-sass', '~> 5.15.1'
+gem 'mysql2'
+gem 'font-awesome-sass'
 gem "paperclip"
-gem 'popper_js', '>= 1.14.3', '< 2'
+gem 'popper_js'
 gem "rolify"
 gem "cancan"
 gem "select2-rails"
@@ -31,16 +31,14 @@ gem 'bootstrap4-datetime-picker-rails'
 gem "cocoon"
 gem 'bootstrap-toggle-rails'
 gem 'wicked'
-gem 'carrierwave', '~> 1.0'
-gem 'will_paginate', '~> 3.1.0'
+gem 'carrierwave'
+gem 'will_paginate'
 gem 'ransack'
 gem "switch_user"
 gem 'state_machines-activerecord'
 gem 'wkhtmltopdf-binary'
 gem 'wicked_pdf'
 gem 'toastr-rails'
-gem 'rails_email_preview'
-gem 'bigdecimal', '1.3.5'
 gem 'onesignal'
 
 gem 'fullcalendar-rails'
@@ -58,7 +56,7 @@ group :development do
   gem 'rb-fsevent', :require=>false
   gem 'rb-inotify', :require=>false
   gem 'spring-commands-rspec'
-  gem 'quiet_assets'
+  gem 'test-unit'
 end
 
 group :production do
