@@ -1,6 +1,6 @@
 class IntranetController < ApplicationController
 
-  before_filter :set_intranet_structure
+  before_action :set_intranet_structure
 
   def set_intranet_structure
     if request.subdomain && request.subdomain != '' && request.subdomain != 'admin'
