@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
 
-  before_filter :check_current_user
+  before_action :check_current_user
 
   def index
     @q = params['search']
