@@ -3,6 +3,7 @@ require File.expand_path('../boot', __FILE__)
 require 'csv'
 require 'rails/all'
 
+
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
@@ -21,7 +22,7 @@ module Funadf
 
     config.quiet_assets = true
 
-    config.active_record.raise_in_transactional_callbacks = true
+    # config.active_record.raise_in_transactional_callbacks = true
     config.assets.precompile += %w( intranet.js )
   end
 end
