@@ -1,6 +1,6 @@
 Rails.application.routes.draw do
 
-  mount RailsEmailPreview::Engine, at: 'emails'
+
   devise_for :users, controllers: { invitations: 'users/invitations' }
 
 
